Delete generator a lot to do but base methods are set

+++ src/Dapper-Layers-Generator/Dapper-Layers-Generator.Core/Generators/GeneratorRepoMain.cs
@@ -180,7 +180,7 @@ using {_settings.TargetNamespaceForDbContext};
             }
 
             //Delete
-            if (TableSettings.DeleteGenerator)
+            if (TableSettings.DeleteGenerator && !string.IsNullOrEmpty(GetPkMemberNamesString()))
             {
                 output.Append($"{tab}{tab}Task DeleteAsync({GetPkMemberNamesStringAndType()});");
